Migrated test to use factory boy
